move backoff settings, send email, and create client to config
